+++ adapters/bevy/client/src/systems.rs
@@ -145,8 +145,8 @@ pub fn translate_world_events<T: Send + Sync + 'static>(world: &mut World) {
                 let mut event_writer = world
                     .get_resource_mut::<Messages<bevy_events::DisconnectEvent<T>>>()
                     .unwrap();
-                for (_, reason) in events.read::<naia_events::DisconnectEvent>() {
-                    event_writer.write(bevy_events::DisconnectEvent::<T>::new(reason));
+                for (_, reason, message) in events.read::<naia_events::DisconnectEvent>() {
+                    event_writer.write(bevy_events::DisconnectEvent::<T>::new(reason, message));
                 }
             }